some work done that I don't wanna lose

+++ models/v1/fake/query_filter.go
@@ -11,10 +11,10 @@ func BuildFleshedOutQueryFilter() *models.QueryFilter {
 	return &models.QueryFilter{
 		Page:          10,
 		Limit:         20,
-		CreatedAfter:  uint64(fake.Date().Unix()),
-		CreatedBefore: uint64(fake.Date().Unix()),
-		UpdatedAfter:  uint64(fake.Date().Unix()),
-		UpdatedBefore: uint64(fake.Date().Unix()),
+		CreatedAfter:  uint64(uint32(fake.Date().Unix())),
+		CreatedBefore: uint64(uint32(fake.Date().Unix())),
+		UpdatedAfter:  uint64(uint32(fake.Date().Unix())),
+		UpdatedBefore: uint64(uint32(fake.Date().Unix())),
 		SortBy:        models.SortAscending,
 	}
 }
